fix: 修改接口类型，处理无法设置属性值的问题
原类型错误而无法设置并处理属性

Log: 安全更新的功能缺失
Task: https://pms.uniontech.com/bug-view-173629.html
Influence: 控制中心-更新

+++ src/plugin-update/operation/updatedbusproxy.h
@@ -51,7 +51,7 @@ public:
 
     Q_PROPERTY(qulonglong UpdateMode READ updateMode WRITE setUpdateMode NOTIFY UpdateModeChanged)
     uint updateMode();
-    void setUpdateMode(uint value);
+    void setUpdateMode(qulonglong value);
 
     Q_PROPERTY(QList<QDBusObjectPath> JobList READ jobList NOTIFY JobListChanged)
     QList<QDBusObjectPath> jobList();
